requirements.txt
django==5.2.12
python-dotenv==1.0.1

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1

# Password hashing
argon2-cffi==23.1.0

# Filtering
django-filter==24.3

# OpenAPI documentation
drf-spectacular==0.27.2